Move test to params/protocol_params_test.go
